requirements.txt
face-recognition==1.3.0
face-recognition-models==0.3.0
Pillow==8.2.0
